+++ src/components/Breadcrumb/Breadcrumb.stories.tsx
@@ -2,12 +2,13 @@ import { faFileAlt, faFolder, faHome } from "@fortawesome/free-solid-svg-icons";
 import { Meta, StoryObj } from "@storybook/react";
 import { generateHiddenArgTypes } from "../../internalUtils/generateHiddenArgTypes";
 import { Breadcrumb } from "./Breadcrumb";
-import { BreadcrumbProps, Crumb } from "./types";
+import { BreadcrumbProps, CrumbType } from "./types";
 
-export const crumbs: Crumb[] = [
+export const crumbs: CrumbType[] = [
   { id: 1, label: "Home", path: "/", icon: faHome },
   { id: 2, label: "Folder", path: "/folder", icon: faFolder },
-  { id: 3, label: "File", path: "/folder/file", icon: faFileAlt },
+  { id: 3, label: "Sub Folder", path: "/folder/subFolder", icon: faFolder },
+  { id: 4, label: "File", path: "/folder/subFolder/file", icon: faFileAlt },
 ];
 
 const meta: Meta<BreadcrumbProps> = {
@@ -15,7 +16,7 @@ const meta: Meta<BreadcrumbProps> = {
   component: Breadcrumb,
   args: { crumbs },
   argTypes: {
-    ...generateHiddenArgTypes(["onClick"]),
+    ...generateHiddenArgTypes(["onClick", "className"]),
     variant: {
       control: {
         type: "radio",