+++ src/components/Form/index.ts
@@ -1,2 +1,3 @@
 export { default as Form } from "./Form";
+export { useForm } from "./hooks/useForm";
 export * from "./types";