+++ src/components/Miscellaneous/index.ts
@@ -1,6 +1,7 @@
 export * from "./ColorPreviewSquare";
 export * from "./HueSliderCanvas";
 export * from "./Icon";
+export * from "./IconPreviewSquare";
 export * from "./Overlay";
 export * from "./ProgressSpinner";
 export * from "./Tooltip";